refactor: split ResumeBuilder.jsx into step components
- Split 1999-line monolith into pages/ResumeBuilder/ with index.jsx
  orchestrator + 7 lazy-loaded step components
- Extracted form state into useResumeWizard, derived/score state into
  useResumeInsights
- Added CustomSectionsStep + useResumeAutosave (localStorage draft
  persistence with restore)
- index.jsx is 199 LOC, each step is its own lazy-loaded chunk

Fixes #4258
