+++ frontend/src/pages/Dashboard.jsx
@@ -537,17 +537,6 @@ export default function Dashboard() {
   <ResumeIndustryOptimizer />
 </motion.div>
 
-  <PortfolioGrammarAssistant />
-</motion.div>
-
-<motion.div variants={itemVariants} className="mb-10">
-  <PortfolioGrammarAssistant />
-</motion.div>
-
-<motion.div variants={itemVariants} className="mb-10">
-  <ResumeBulletEnhancer />
-</motion.div>
-
 <div className="grid lg:grid-cols-2 gap-10"></div>
 
             <div className="grid lg:grid-cols-2 gap-10">
@@ -701,4 +690,3 @@ export default function Dashboard() {
     </div>
   )
 }
-
